+++ src/pickers/index.tsx
@@ -2,8 +2,7 @@
 https://github.com/deseretdigital/dayzed/pull/25
 He didn't publish the components to npm, so I copied his work */
 
-import BaseDatePicker from './base';
-import DatePicker from './basic';
+import BasicDatePicker from './basic';
 import RangeDatePicker from './range';
 
-export { BaseDatePicker, DatePicker, RangeDatePicker };
+export { BasicDatePicker, RangeDatePicker };
